Add files via upload
This design of this system is web-based type, so the user also can directly use the system by connect to internet The users of this system are given to three groups, which are administrators, teachers and parents. The functions of administrators and teachers are read, write and edit, while the parents their functions are only read and write. Through this system, administrators and teachers can manage all student data and information easily' and the parents also can easily know their children performance easily.